correction on migration script

+++ data/migrations/2.5.5to2.6.0.sql
@@ -247,7 +247,7 @@ BEGIN;
     FOR EACH STATEMENT
     EXECUTE PROCEDURE gn_synthese.fct_trig_insert_in_cor_area_synthese_on_each_statement();
 
-
+    DROP TRIGGER tri_update_cor_area_synthese ON gn_synthese.synthese;
     CREATE TRIGGER tri_update_cor_area_synthese
     AFTER UPDATE OF the_geom_local, the_geom_4326 ON gn_synthese.synthese
     FOR EACH ROW
@@ -313,35 +313,45 @@ BEGIN;
 
 
     -- Update VIEW export OCCHAB
-    DROP VIEW IF EXISTS pr_occhab.v_export_sinp;
-    CREATE VIEW pr_occhab.v_export_sinp AS
-    SELECT 
-    s.id_station,
-    s.id_dataset,
-    s.id_digitiser,
-    s.unique_id_sinp_station as "identifiantStaSINP",
-    ds.unique_dataset_id as "metadonneeId",
-    nom1.cd_nomenclature as "dSPublique",
-    to_char(s.date_min, 'DD/MM/YYYY'::text)as "dateDebut",
-    to_char(s.date_max, 'DD/MM/YYYY'::text)as "dateFin",
-    s.observers_txt as "observateur",
-    nom2.cd_nomenclature as "methodeCalculSurface",
-    public.st_astext(s.geom_4326) as "geometry", -- Pourquoi rajouter st_astext?
-    public.st_asgeojson(s.geom_4326) as geojson,
-    s.geom_local,
-    nom3.cd_nomenclature as "natureObjetGeo",
-    h.unique_id_sinp_hab as "identifiantHabSINP",
-    h.nom_cite as "nomCite",
-    h.cd_hab as "cdHab",
-    h.technical_precision as "precisionTechnique"
-    FROM pr_occhab.t_stations as s
-    JOIN pr_occhab.t_habitats h on h.id_station = s.id_station
-    JOIN gn_meta.t_datasets ds on ds.id_dataset = s.id_dataset
-    LEFT join ref_nomenclatures.t_nomenclatures nom1 on nom1.id_nomenclature = ds.id_nomenclature_data_origin
-    LEFT join ref_nomenclatures.t_nomenclatures nom2 on nom2.id_nomenclature = s.id_nomenclature_area_surface_calculation
-    LEFT join ref_nomenclatures.t_nomenclatures nom3 on nom3.id_nomenclature = s.id_nomenclature_geographic_object
-    LEFT join ref_nomenclatures.t_nomenclatures nom4 on nom4.id_nomenclature = h.id_nomenclature_collection_technique;
+    DO $$
+      BEGIN
+      IF EXISTS (SELECT * FROM pg_catalog.pg_namespace WHERE nspname = 'pr_occhab')
+        THEN 
+          DROP VIEW IF EXISTS pr_occhab.v_export_sinp;
+          CREATE VIEW pr_occhab.v_export_sinp AS
+            SELECT 
+            s.id_station,
+            s.id_dataset,
+            s.id_digitiser,
+            s.unique_id_sinp_station as "identifiantStaSINP",
+            ds.unique_dataset_id as "metadonneeId",
+            nom1.cd_nomenclature as "dSPublique",
+            to_char(s.date_min, 'DD/MM/YYYY'::text)as "dateDebut",
+            to_char(s.date_max, 'DD/MM/YYYY'::text)as "dateFin",
+            s.observers_txt as "observateur",
+            nom2.cd_nomenclature as "methodeCalculSurface",
+            public.st_astext(s.geom_4326) as "geometry", -- Pourquoi rajouter st_astext?
+            public.st_asgeojson(s.geom_4326) as geojson,
+            s.geom_local,
+            nom3.cd_nomenclature as "natureObjetGeo",
+            h.unique_id_sinp_hab as "identifiantHabSINP",
+            h.nom_cite as "nomCite",
+            h.cd_hab as "cdHab",
+            h.technical_precision as "precisionTechnique"
+            FROM pr_occhab.t_stations as s
+            JOIN pr_occhab.t_habitats h on h.id_station = s.id_station
+            JOIN gn_meta.t_datasets ds on ds.id_dataset = s.id_dataset
+            LEFT join ref_nomenclatures.t_nomenclatures nom1 on nom1.id_nomenclature = ds.id_nomenclature_data_origin
+            LEFT join ref_nomenclatures.t_nomenclatures nom2 on nom2.id_nomenclature = s.id_nomenclature_area_surface_calculation
+            LEFT join ref_nomenclatures.t_nomenclatures nom3 on nom3.id_nomenclature = s.id_nomenclature_geographic_object
+            LEFT join ref_nomenclatures.t_nomenclatures nom4 on nom4.id_nomenclature = h.id_nomenclature_collection_technique;
 
+        ELSE
+         RAISE NOTICE 'Schema pr_occhab does not exists';
+      END IF;
+    END$$;
+
+    
 
     -- Révision de la vue validation de gn_commons pour disposer du nom français et latin
     DROP VIEW gn_commons.v_synthese_validation_forwebapp;
